belle_croissant/requirements.txt
asgiref==3.11.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
